Small style changes
I want to keep the writers' styles as much as possible, while providing consistency across the whole page/site. Obviously much of the style argument is going to be personal preference, as we come from many different backgrounds, and even the academics can't agree. For the descriptions of everyone, I've made them all into standard complete sentences. The Oxford comma will be applied across the board (though this is probably up for debate).