New section and, website, grammar question
Added Goods after discussion with fluffypony
Added cryptonit.net under this section
Used the esperanto form of monero. This has implication, because so far we avoided the matter of choosing because we avoided the plural. This doesn't mean only English plural or only Esperanto plural is accepted - both could, like both color and colours are accepted. Still, style-wise, it is better to stick to one word (color everywhere or colour everywhere) so this question still remains for -s vs -j.