Create unofficial MoneroLink response
I do not have the images attached (sorry), but I would like to post the paper I have written here to the Monero blog: https://1drv.ms/w/s!AjOt8D-0YjBHgYg_onISH13gCSfKng

My apologies for not knowing much of the formatting. Perhaps someone can help me here? I promise I will improve for subsequent proposals :)

I removed the pre-preface for 2 reasons: 1) a pre-preface is poor form in general, and 2), if published on the GetMoneor website, the paper should be a more formal response, even if not official

Images are not attached. My apologies for not including them in the actual formatting.