Update index.html
Monero is secure
Do we really want the first sentence about the security of Monero to include "extremely difficult to hack and steal your funds. "  To me that doesn't sound safe for large amounts if it's only "extremely difficult"... and why mention "hack" and "steal" at all?  Further, if you are trying to define "decentralized cryptocurrency" the definition is absolutely NOT "extremely difficult to hack and steal your funds.".  You could say the same of cash in a bank vault, an ATM machine, a centralized currency, or the safe in my basement...  Those are all "extremely difficult to hack and steal" from.  It seems to me we should be explaining what decentralized cryptocurrency is and perhaps stating it's the safest method known to mankind to store value, outside of Fort Knox.  Not that its hard to hack and steal!


Monero is fungible - if we're going to say it's untraceable above, we may as well remove the whole "In its current state, it is extremely unlikely that Monero will ever be..." and just bluntly state that it will never be blacklisted due to previous association.  Also, very few people know what fungible even means, so I have added the definition so people know what they're reading.