Core team post: remove extraneous globalism rhetoric
Monero is a software. Movements are centralized bodies which have
leaders and funders - many of whom are working behind the scenes to
spread their agenda. Monero is not a movement anymore than Windows is a
movement (although Windows is a bowel movement, that's for sure).

Globalism is not a requisite for privacy and, ironically, globalism is the
antithesis of decentralization as it requires being built on centralized
structures: both physically, socially, and metaphorically. Monero can
scale down to a small, local community and still serve it's purpose.