Add possibility to play localized videos + welcome video in Russian
This commit add the possibility to add videos with the voice-over in different languages.
The videos should be uploaded in /media/LANG/, where lang is the language code, and should
have the same name and extension of the original in English.

This commit also adds the "Monero Promo" video with voice-over in Russian, made by the
xmr.ru community.