Remove 'langs' from language files and move them to their own 'languages.yml' file in /_data
The list of languages doesn't need to be translated, so there is no reason to keep it in the language files. Moved to its own
file (/_data/languages.yml). Now the language dropdown parses the languages from there.
This change is also needed for the migration to Weblate (which would ask translators to translate the language names).