Meta tags: Change opengraph logo and add possiblity to use custom image
The old logo showing on social medias is too large and end up being trimmed. Replaced it with the simple Monero Logo.
If an image is specified using 'image:' in the front matter of a blog post, jekyll will show that image instead. This also allow us to easily
make specific meta-images for each blog posts. I will add the possibility to add the image in the body of the post in a
subsequent PR.