Improve RSS experience (install 'jekyll-feed' plugin)
- Use 'jekyll-feed' plugin to manage RSS feed and remove existing feed.xml file
- Add feed image in main blog tab ('All posts')

The 'jekyll-feed' plugin makes possible for aggregators to automatically detect the existence of a feed (feed_meta tag in header)
and creates a 'feed.atom' file containing the last 20 blog posts and locates it in the root directory. The structure is compatible
with our current system, including the possibility to show a picture if it's included in the post (#1002).