Update downloads.yml
Incognito Wallet is an open source mobile wallet app for iOS and Android devices that supports XMR. The wallet designed and developed by the Incognito Dev team. The purpose this wallet is providing an option of storing XMR funds private. Following the requirements listed at https://repo.getmonero.org/monero-project/monero-site/-/issues/982 we are asking if we the wallet can be listed under the Downloads > Mobile and Light Wallets section of the https://www.getmonero.org/