Add Swapzone to merchant's index
Hey, team!
We are an aggregator of crypto exchanges that have already won the attention of the community. We work with trusted partners and will be glad to have the opportunity to be added to your website :) We are constantly working with XMR on Reddit, and we believe that anonymity is everything!