remove Cake wallet
This is a principled removal, on the grounds that the Cake wallet people
are squatting the monero.com domain to push a version of Cake wallet
they call the "monero.com" wallet.

When called out, representatives of Cake wallet claimed that is is not
squatting, despite the fact $project.com is the most canonical domain
name possible for a project, and that they are trustworthy, therefore
it's fine for them to have that domain (they're a company, the bottom
line is what counts, and even if the current owners do believe this,
any buyers would turn against us without a second thought). Moreover,
they had the audacity to claim I did not complain early enough, as if
I was suposed to be keeping track of who owned monero themed domains.
Last, they claim that they do mention on the website that it is not
the official monero domain. While true, close to nobody reads the web
page footer, and they know it full well.

The intent seems glaringly obvious: to dishonestly cause visitors to
get their software by making it seem, at least for those people who
do not look really hard, that they are the monero project.

Being a company which considers its bottom line, not earning the money
they expected to get from this unethical move is likely the only thing
that will get them to reconsider any further similar moves, and their
wallet being removed from the monero site will help with this.

AFAIK the previous owner of monero.com was some building company, and
their registration predated our project, so there was no possible
confusion on the part of visitors.

Having worked for the Monero project for YEARS of my life, I feel pretty
aggrieved by this dishonesty, and thus I am registering my protest
by this patch, which I acknowledge has little change of being merged,
since practicalities will likely outweigh ethics. However, I tend to
weigh ethics more than practicalities, hence my work on monero for
all these years since pretty much its inception.