Add ZENX to Swappers
Adds ZENX, a privacy-first, no-JavaScript, stateless Monero swap aggregator, to the Swappers section of the Merchants & Exchanges page for future consideration once it meets the community criteria. ZENX is actively operating and listed on other community resources like kycnot.me and Monerica. This PR is to get it on record and ready for inclusion when eligible. https://zenx.ink